+++ packages/core/src/bar-chart.ts
@@ -146,8 +146,8 @@ export class BarChart extends BaseAxisChart {
 						.attr("y", d => this.y(Math.max(0, d.value)))
 						.attr("width", this.x1.bandwidth())
 						.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
-						.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
-						.attr("stroke", d => this.options.accessibility ? this.colorScale[d.datasetLabel](d.label) : null)
+						.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
+						.attr("stroke", d => this.options.accessibility ? this.getStrokeColor(d.datasetLabel, d.label, d.value) : null)
 						.attr("stroke-width", Configuration.bars.default.strokeWidth)
 						.attr("stroke-opacity", d => this.options.accessibility ? 1 : 0)
 						// Make the data element accessible and pass x and y values for the data point, and the dataset label
@@ -195,9 +195,9 @@ export class BarChart extends BaseAxisChart {
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
 			.style("opacity", 0)
 			.transition(this.getFillTransition())
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
 			.style("opacity", 1)
-			.attr("stroke", (d: any) => this.colorScale[d.datasetLabel](d.label))
+			.attr("stroke", (d: any) => this.getStrokeColor(d.datasetLabel, d.label, d.value))
 			.attr("stroke-width", Configuration.bars.default.strokeWidth);
 
 		addedBars.selectAll("rect.bar")
@@ -211,9 +211,9 @@ export class BarChart extends BaseAxisChart {
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
 			.style("opacity", 0)
 			.transition(this.getFillTransition())
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
 			.style("opacity", 1)
-			.attr("stroke", (d: any) => this.colorScale[d.datasetLabel](d.label))
+			.attr("stroke", (d: any) => this.getStrokeColor(d.datasetLabel, d.label, d.value))
 			.attr("stroke-width", Configuration.bars.default.strokeWidth);
 
 		// Remove bar groups are no longer needed
@@ -264,8 +264,8 @@ export class BarChart extends BaseAxisChart {
 			.attr("y", d => this.y(Math.max(0, d.value)))
 			.attr("width", this.x1.bandwidth())
 			.attr("height", d => Math.abs(this.y(d.value) - this.y(0)))
-			.attr("fill", d => this.getFillScale()[d.datasetLabel](d.label))
-			.attr("stroke", d => this.options.accessibility ? this.colorScale[d.datasetLabel](d.label) : null);
+			.attr("fill", d => this.getFillColor(d.datasetLabel, d.label, d.value))
+			.attr("stroke", d => this.options.accessibility ? this.getStrokeColor(d.datasetLabel, d.label, d.value) : null);
 	}
 
 	resizeChart() {
@@ -304,7 +304,7 @@ export class BarChart extends BaseAxisChart {
 			.on("mouseover", function(d) {
 				select(this)
 					.attr("stroke-width", Configuration.bars.mouseover.strokeWidth)
-					.attr("stroke", self.colorScale[d.datasetLabel](d.label))
+					.attr("stroke", self.getStrokeColor(d.datasetLabel, d.label, d.value))
 					.attr("stroke-opacity", Configuration.bars.mouseover.strokeOpacity);
 
 				self.showTooltip(d, this);
@@ -321,7 +321,7 @@ export class BarChart extends BaseAxisChart {
 				const { strokeWidth, strokeWidthAccessible } = Configuration.bars.mouseout;
 				select(this)
 					.attr("stroke-width", accessibility ? strokeWidthAccessible : strokeWidth)
-					.attr("stroke", accessibility ? self.colorScale[d.datasetLabel](d.label) : "none")
+					.attr("stroke", accessibility ? self.getStrokeColor(d.datasetLabel, d.label, d.value) : "none")
 					.attr("stroke-opacity", Configuration.bars.mouseout.strokeOpacity);
 
 				self.hideTooltip();